Add circle tests case for polynomial (quadratic) kernel
Two test scenarios: one where the circle is centered on the origin. The
quadratic kernel allows to learn this model perfectly, hence a hard
margin (small nu) can be used. The other is a circle that is offset from the
center. The quad kernel cannot learn this model, as its decision boundary is a
conic section centered at the origin. The learnt model will produce an eccentric
ellipse that is also inclined to capture most of the offset circle, but there
will be some misclassifications. Thus, a much softer margin (large nu) is
required and one cannot expect close-to-perfect success rates.