Make model::operator() return all decision function values
Return type is pair of resulting label and an appropriately sized std::array
containing all decision function values, the number of which is exposed as
model::nr_classifiers. A special case is binary classification with only one
decision function: the array is replaced with just a plain double for simplicity
and to keep backward-compatibility. The type of the decision values is exposed
as model::decision_type.