Models do not longer serve as classifier in biclassification
Revised the idea introduced in 619b789: svm::model no longer implements the
Classifier concept in the case of biclassification. The reason is consistent
ownership: introspectors previously referenced models, but would own
classifiers. Since models are not copyable, they would have to be treated
differently in introspectors, complicating those and leading to code duping.

New approach: models are only used to generate classifiers through the
classifier member function. The label arguments may be omitted for
biclassification. The classifier then internally references the model, can be
copied and iterated over. The model is no longer iterable.

This BREAKS introspectors for now and will be fixed down the line.