Classifiers are callable, know their rho(), and yield params
Moved some capabilities of models to classifiers. The classifiers now are
callable and return a pair of label and decision function value. They thus serve
as a drop-in replacement for biclassification models (same return type). Note
that it is still possible *and more efficient* to call the model, if decision
function values for *all* classifiers are required.

The rho parameter is now a property of the classifier; appropriate value will be
picked out from the array. This effectively hides implementation details of
LIBSVM from the users of the wrapper. o/

Classifiers also get a params() member function because introspectors might want
access to the kernel parameters. (In preparation of introspector refit.)