Revive model::rho() function & treat permutation (+signs) correctly
The model::classifier_type::rho() function previously did not account for the
sign of the classifier correctly, which has been fixed.

The model::rho() had been previously removed in 44e16c which has been reverted.
In the case of biclassification, it returns a single double, in agreement with
the original behavior. For multi-classification, an array is returned which is
permuted correctly to match the ones returned by the call operator and
model::classifiers().