Use std::array as classifier_arr_t, rather than std::vector
Previously, std::vector was used because of difficulty of constructing
classifier_type objects (which are not default-constructible!) when initializing
the array. (With vector, this could be circumvented by emplacing classifiers;
something not possible with std::array). This new version solves this problem by
using std::integer_sequence and variadic expansion to construct the classifiers
all at once when initializing the array.