Fix/workaround bug: capture in pack expansion in classifiers_impl
Not sure if this is a compiler bug or a bug on my part. Anyway, there seem to be
issues (at least with GCC 7.1) when using lambdas with captures inside variadic
pack expansions, as is the case in model::classifiers_impl. `this` was only
correctly captured in the first lambda. Worked around the issue by not using
IILE but defining the lambda before the pack expansion and invoking it there.

Confer: https://stackoverflow.com/questions/39128413